Retheme the whole app to Outfit, softer surfaces, rounder corners
Jack picked direction 04 off the mockups -- "i like this more... but the
same green we are already using.. and make this theme universal across all
tabs" -- and separately: "I feel like these are too cluttered."

Universal comes free because nothing outside :root names a typeface. Two
token lines retheme every tab; the only non-token font references in the
codebase turned out to be the JetBrains Mono @import and two comments.
Verified by computed style in the browser rather than by reading the CSS:
body and .h1 both resolve to Outfit, and the only families the page loads
are Outfit, Phosphor and JetBrains Mono.

  font        Inter -> Outfit, everywhere
  title       800 / 21px -> 600 / 26px, tracking -0.025em
  surfaces    #0b0c11 / #131419 / #191b22 -> #0c0f12 / #171c21 / #1e242b
  radii       6/14/18/22 -> 10/20/24/28, buttons 10 -> 16
  accent      unchanged, #4ce08f

Lifting the surfaces a step is what makes the bigger radii read as soft
rather than as holes cut in a black page.

The clutter was the copy, not the chrome. The mockup he liked also has an
uppercase kicker; the difference is that friendlyDate produced "Tuesday,
September 15", so the kicker ran nearly thirty characters of wide-tracked
uppercase above a six-character title. Now "Tue, Sep 15 · Week 1" -- same
information, half the width, and it lands on all three day views at once
since they share the one formatter.

Two things the token swap could not reach, both found by looking at the
rendered page:

  - Four hardcoded hexes tuned to the old near-black: the mobile and
    desktop header gradients, which would have sat darker than the
    surfaces above them.
  - The JACKED lockup asks for weight 800 and .hero-value for 900, but the
    font link only carried 300-700, so the browser faked both by smearing
    the 700 outline. On a 40px uppercase wordmark that reads as a
    different, wider typeface -- I mistook it for one. Both real weights
    are loaded now.

Verified: oxlint clean of error-level rules, tsc -b silent, vite build ok;
tests 268, 0 failing. Landing screen checked in the browser at 375x812 --
the only screen reachable without a session, so the signed-in screens are
unverified.

+++ src/data/dayNumbering.ts
@@ -16,12 +16,16 @@ export function dayDisplayTitle(day: Pick<TrainingDay, "code">): string {
   return `Day ${dayNumberInWeek(day)}`;
 }
 
-/** "Friday, September 4" -- the one date format every day view uses in its kicker, so today's session, an
- * upcoming preview, and a logged day all read the same way. Today's used to show set progress here
- * instead ("Week 1 · 0 of 15 sets"), which made it the odd one out; the remaining-set count is already
- * shown above the Finish button where it's actually actionable. */
+/** "Fri, Sep 4" -- the one date format every day view uses in its kicker, so today's session, an upcoming
+ * preview, and a logged day all read the same way. Today's used to show set progress here instead
+ * ("Week 1 · 0 of 15 sets"), which made it the odd one out; the remaining-set count is already shown above
+ * the Finish button where it's actually actionable.
+ *
+ * Abbreviated because the kicker is set in uppercase with wide tracking, and the long form made it the
+ * loudest thing on the screen -- "TUESDAY, SEPTEMBER 15 · WEEK 1" ran nearly thirty characters above a
+ * six-character title. Jack: "I feel like these are too cluttered." Same information, half the width. */
 export function friendlyDate(iso: string): string {
-  return new Date(`${iso}T00:00:00`).toLocaleDateString("en-US", { weekday: "long", month: "long", day: "numeric" });
+  return new Date(`${iso}T00:00:00`).toLocaleDateString("en-US", { weekday: "short", month: "short", day: "numeric" });
 }
 
 /** The shared kicker for every day view: "Friday, September 4 · Week 1". */
